Complete rewrite.
Content is based on my own knowledge, as well as a review of interesting
security issues in the news over the last month.
Restructuring and rewording is based on review by a non-technical
friend, including a final review.

A lot of this restructure is intended to make the guide more widely
useful. By splitting in to "Basic" and "Advanced" sections we're hoping
that this will get handed around among more "new" users. Many of these
users will likely skim the "Advanced" section, raising awareness. Also
hopefully it'll get passed around virally.